pyproject.toml
[project]
name = "insurance-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.4",
    "dash>=3.3.0",
    "dash-bootstrap-components>=2.0.4",
    "gunicorn>=23.0.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "s3fs>=0.4.2",
]
